requirement.txt
streamlit
pillow
openai
cromadb
requests